requirements.txt
google-search-results
python-dotenv
serpapi